+++ src/components/Body/index.js
@@ -1,5 +1,5 @@
 export { default as Body } from "./Body";
 export { default as Content } from "./Content";
 export { default as SideBar } from "./SideBar";
-export { default as UserProfilePic } from "./UserProfilePic";
+export { default as UserCard } from "./UserCard";
 export { default as NavCard } from "./NavCard";
